backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
PyJWT==2.10.1
bcrypt==4.3.0
python-dotenv==1.1.1

